biometric_multimodal/requirements.txt
flask==3.0.2
flask-cors==4.0.0
PyQt6==6.10.2
opencv-python==4.13.0.92
numpy==2.4.3
scipy==1.16.1
scikit-learn==1.8.0
scikit-image==0.26.0
Pillow==11.3.0
PyYAML==6.0.3
tqdm==4.67.1
joblib==1.5.2
torch==2.10.0
torchvision==0.25.0
timm==1.0.22
onnxruntime==1.24.4
matplotlib==3.10.7
kagglehub==0.3.12
face-recognition==1.3.0
